+++ app/src/main/java/com/example/bakedeggs/AddContact/AddFragment.kt
@@ -1,0 +1,61 @@
+package com.example.bakedeggs.AddContact
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.bakedeggs.R
+import com.example.bakedeggs.databinding.FragmentAddBinding
+
+private const val CONTACT_DATA = "param1"
+
+class AddFragment : Fragment() {
+    private var _binding : FragmentAddBinding? = null
+    private val binding  get() = _binding!!
+
+    private var param1: String? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.let {
+            param1 = it.getString(CONTACT_DATA)
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentAddBinding.inflate(inflater,container,false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        //binding.
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+    companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @return A new instance of fragment AddFragment.
+         */
+        // TODO: Rename and change types and number of parameters
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            AddFragment().apply {
+                arguments = Bundle().apply {
+                    putString(CONTACT_DATA, param1)
+                }
+            }
+    }
+}
