+++ app/src/main/res/layout/fragment_add.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddFragment">
+    tools:context=".AddContact.AddFragment">
 
     <ScrollView
         android:layout_width="match_parent"
